review: Move all classes in statefun-smoke-e2e-java to a more specific package
