+++ statefun-e2e-tests/statefun-smoke-e2e-java/pom.xml
@@ -53,7 +53,7 @@ under the License.
             <scope>test</scope>
         </dependency>
 
-        <!-- smoke end-to-end common -->
+        <!-- smoke end-to-end driver -->
         <dependency>
             <groupId>org.apache.flink</groupId>
             <artifactId>statefun-smoke-e2e-driver</artifactId>
@@ -76,7 +76,7 @@ under the License.
                     </descriptorRefs>
                     <archive>
                         <manifest>
-                            <mainClass>org.apache.flink.statefun.e2e.smoke.CommandInterpreterAppServer</mainClass>
+                            <mainClass>org.apache.flink.statefun.e2e.smoke.java.CommandInterpreterAppServer</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
@@ -87,6 +87,31 @@ under the License.
                         <goals>
                             <goal>single</goal>
                         </goals>
+                    </execution>
+                </executions>
+            </plugin>
+
+            <!-- Generate the Command messages -->
+            <plugin>
+                <groupId>com.github.os72</groupId>
+                <artifactId>protoc-jar-maven-plugin</artifactId>
+                <version>${protoc-jar-maven-plugin.version}</version>
+                <executions>
+                    <execution>
+                        <id>generate-protobuf-sources</id>
+                        <phase>generate-sources</phase>
+                        <goals>
+                            <goal>run</goal>
+                        </goals>
+                        <configuration>
+                            <includeStdTypes>true</includeStdTypes>
+                            <protocVersion>${protobuf.version}</protocVersion>
+                            <cleanOutputFolder>true</cleanOutputFolder>
+                            <inputDirectories>
+                                <inputDirectory>src/main/protobuf</inputDirectory>
+                            </inputDirectories>
+                            <outputDirectory>${basedir}/target/generated-sources/protoc-jar</outputDirectory>
+                        </configuration>
                     </execution>
                 </executions>
             </plugin>
