+++ statefun-e2e-tests/statefun-smoke-e2e-java/src/main/java/org/apache/flink/statefun/e2e/smoke/java/CommandInterpreter.java
@@ -16,9 +16,9 @@
  * limitations under the License.
  */
 
-package org.apache.flink.statefun.e2e.smoke;
+package org.apache.flink.statefun.e2e.smoke.java;
 
-import static org.apache.flink.statefun.e2e.smoke.Constants.*;
+import static org.apache.flink.statefun.e2e.smoke.java.Constants.*;
 
 import java.time.Duration;
 import java.util.Objects;
